把 reserve ui 交互的部分抽离出来
